requirements.txt
asgiref==3.8.1
certifi==2024.6.2
Django==5.0.6
django-taggit==5.0.1
django-widget-tweaks==1.5.0
pillow==10.3.0
python-decouple==3.8
sqlparse==0.5.0
